Refactor: Move firstline filtering to filter module
To enable the HIRS FCDR (which depends on typhon) to implement its own
overlap filtering without difficult subclassing, refactor overlap
filtering to classes in the filter module, based on the compasition over
inheritance principle.  Then FCDR_HIRS can fix
https://github.com/FIDUCEO/FCDR_HIRS/issues/7 either by adding an
implementation here on within FCDR_HIRS.
